# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Gearbox API",
        "description": "REST API untuk Sistem Booking & Manajemen Layanan Bengkel Mobil",
        "contact": {
            "email": "admin@gearbox.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/api/services": {
            "get": {
                "tags": [
                    "Services"
                ],
                "summary": "Get semua layanan",
                "operationId": "97de511018def32ca0ccbc800756c538",
                "parameters": [
                    {
                        "name": "category",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "mesin",
                                "kelistrikan",
                                "bodi",
                                "ac"
                            ]
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "active",
                                "inactive"
                            ]
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List layanan dengan pagination"
                    }
                }
            }
        },
        "/api/services/{id}": {
            "get": {
                "tags": [
                    "Services"
                ],
                "summary": "Get detail layanan",
                "operationId": "dce5b306a141bdec91eff8a15534d7d7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detail layanan"
                    },
                    "404": {
                        "description": "Layanan tidak ditemukan"
                    }
                }
            }
        },
        "/api/admin/services": {
            "post": {
                "tags": [
                    "Admin"
                ],
                "summary": "Tambah layanan baru (Admin)",
                "operationId": "0ee39d35a8b7bf7a58228de0ae8ae2f8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "price",
                                    "duration_minutes",
                                    "category"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Ganti Oli Mesin"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Penggantian oli mesin"
                                    },
                                    "price": {
                                        "type": "number",
                                        "example": 150000
                                    },
                                    "duration_minutes": {
                                        "type": "integer",
                                        "example": 30
                                    },
                                    "category": {
                                        "type": "string",
                                        "enum": [
                                            "mesin",
                                            "kelistrikan",
                                            "bodi",
                                            "ac"
                                        ]
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "active",
                                            "inactive"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Layanan berhasil ditambahkan"
                    },
                    "403": {
                        "description": "Unauthorized — bukan admin"
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/services/{id}": {
            "put": {
                "tags": [
                    "Admin"
                ],
                "summary": "Update layanan (Admin)",
                "operationId": "92c769da3b83d63e2f94450cc5ad2f7d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "price": {
                                        "type": "number"
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "active",
                                            "inactive"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Layanan berhasil diupdate"
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Layanan tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Admin"
                ],
                "summary": "Hapus layanan (Admin)",
                "operationId": "16b6b0f50ca2c7104dff94346a89fb65",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Layanan berhasil dihapus"
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Layanan tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Masukkan token dari response login",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}